jitterentropy-rustrngd: cast RNDADDENTROPY to i32 for ioctl on musl/aarch64
libc::ioctl takes i32 on musl aarch64, but RNDADDENTROPY was defined as
c_ulong (u64 on 64-bit). The value 0x40085203 fits in 32 bits; cast at
the call site to match the expected type.

+++ utils/jitterentropy-rustrngd/src/src/main.rs
@@ -108,7 +108,7 @@ unsafe fn inject_entropy(ec: *mut jent_ffi::RandData, fd: i32) -> Result<(), Str
         buf,
     };
 
-    let ret = ioctl(fd, RNDADDENTROPY, &rpi as *const RandPoolInfo);
+    let ret = ioctl(fd, RNDADDENTROPY as i32, &rpi as *const RandPoolInfo);
     if ret < 0 {
         return Err(format!(
             "RNDADDENTROPY ioctl failed: {}",
